requirements.txt
numpy
requests
python-dotenv
tqdm
streamlit